Fixed CarHelperServiceTest.
Some tests were failing because of mHelper was a spy, which is actually
a wrapper. As a result, there were 2 CarServiceHelperInstances, so the
mCarService was set in one but was null in another, causing some
operations to be queued.

This issue was fixed by spliting mHelper into 2 objects (the other
being mHelperSpy), which introduced timing issues (fixed by sleeping
to make sure the HAL response time is not 0).

Bug: 160819994
Test: atest CarHelperServiceTest.

Change-Id: I455a25f14f1b7601ded01c99d10e976f54228e4a