Remove the user assignment to the current user for the overlay display in ExtraDisplayMonitor.
In the automotive multi-user, multi-display environment that supports
background users with UI access on assigned displays (a.k.a. visible
background users), it's not allowed to create a window or launch an
activity on an unassigned display.
Since overlay displays are assigned to the current user, when the
visible background user is running the test, it does not have
permission to launch an activity on the overlay display.
To resolve this, the user assignment to the current user in
ExtraDisplayMonitor is removed, and it is modified so that
the overlay display is assigned to the visible background user during
the creation of the overlay display in the test.

Flag: EXEMPT (bug fix)
Bug: 362325749
Test: atest CtsWindowManagerDeviceActivity:ActivityLifecycleTopResumedStateTests#testFinishOnDifferentDisplay_nonFocused
Test: atest --user-type secondary_user CtsWindowManagerDeviceActivity:ActivityLifecycleTopResumedStateTests#testFinishOnDifferentDisplay_nonFocused
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerDeviceActivity:ActivityLifecycleTopResumedStateTests#testFinishOnDifferentDisplay_nonFocused

Signed-off-by: Dohyun Lee <dohyun.lee@lge.com>
(cherry picked from https://partner-android-review.googlesource.com/q/commit:f22b788db3e303ebd223db73e08e41db7042529c)
Change-Id: I76bf2255fda57df5ebbe0c8fab3b3f0285b216d5