Modify CarLaunchModifier to specify the proper target display.
The activity needs to be available to launch on the target display based on the displayArea of the source activityRecord or the caller context.
I add logic on CarLaunchModifier to handle these parts.
Additionally, I modified logic to check the user assigned to the display to ensure the activity can be launched on the proper display, even if the display is not mapped to an occupantZone.

Test: atest CtsWindowManagerDeviceMultiDisplay
Test: atest CtsWindowManagerDeviceMultiDisplay --user-type secondary_user
Test: atest CtsWindowManagerDeviceMultiDisplay --user-type secondary_user_on_secondary_display
Test: atest CarLaunchParamsModifierUpdatableTest
Bug: 356518965
Flag: EXEMPT bugfix
(cherry picked from https://partner-android-review.googlesource.com/q/commit:ed9c1e835d5ca01e73c0dbeecf39d78798f5c5f0)

Change-Id: I1b50f14fee60fd022c4896c250df47c57ad2ede3