Pass activity launches to ScaleProvider
CarDisplayCompatScaleProviderUpdatableImpl keeps a map of uid to
display, but often wasn't getting uid updates from the
ActivityInterceptor because the interceptor depends on the display
compat host app implementation that is no longer used. Since there
is no host app activity changes would never be propogated to the
scale provider. This caused the provider to often assume apps were
being launched on the default display even when it wasn't.

Bug: 402129197
Test: atest com.android.server.wm.CarDisplayCompatActivityInterceptorTest
Flag: EXEMPT bug_fix

Change-Id: I6126b015254de7f06cc6390f8fda9a35d2d81826