Ensure activity is launched on the right display on AAOS
In AAOS multi-user multi-display enviriment,
when starting an activity for a specific user from Shell
(e.g., via am start --user 10), the activity was incorrectly
launching on the focused display (display 2) instead of
the user's assigned display (display 0).

This was caused by CarLaunchParamsModifier skipping its
display selection logic and deferring to the default
TaskLaunchParamsModifier. This happened because for the
activity launch by the shell, the task and activity objects are null,
so CarLaunchParamsModifier was unable to determine the
target userId.

This resulted in the TaskLaunchParamsModifier's fallback
logic being triggered, which defaults to launching on the
currently focused display.

This change fixed that by exposing the neccessary info from the
underlying ActivityStarter.Request.

Fixes: 429003504
Fixes: 419656704
Test: atest CtsWindowManagerDeviceActivity:android.server.wm.activity.StartActivityAsUserTests#startActivityAsValidUserWithOptions
Flag: EXEMPT BUGFIX

Change-Id: I23b7b57b7c5d20ab32bffbba121851edd707a125